refactor: Type Don't reuse defined used file
