+++ client/src/Component/Core/InputLabel.tsx
@@ -1,8 +1,21 @@
 import React from "react";
 import { css } from "@emotion/react";
 import { Input } from "@Atom/Input/Input";
-import { InputLabelType } from "@Common/type";
 
+export const InputLabel: React.FC<InputLabelType> = ({ refProps, label, placeholder }) => {
+  return (
+    <div css={InputLabelStyle}>
+      <p css={LabelStyle}>{label}</p>
+      <Input placeholder={placeholder} ref={refProps} />
+    </div>
+  );
+};
+
+type InputLabelType = {
+  label: string;
+  placeholder?: string;
+  refProps?: React.MutableRefObject<HTMLInputElement | null>;
+};
 const InputLabelStyle = css`
   margin: 1% 2%;
   width: 300px;
@@ -14,12 +27,3 @@ const LabelStyle = css`
   height: 20%;
   width: 90%;
 `;
-
-export const InputLabel: React.FC<InputLabelType> = ({ refProps, label, placeholder }) => {
-  return (
-    <div css={InputLabelStyle}>
-      <p css={LabelStyle}>{label}</p>
-      <Input placeholder={placeholder} ref={refProps} />
-    </div>
-  );
-};
