+++ client/src/Component/Core/LinkButton.tsx
@@ -1,11 +1,18 @@
-import React from "react";
+import React, { RefObject } from "react";
 import { Link } from "react-router-dom";
 import { Button } from "@Atom/.";
 import { NotReadNum } from "@Molecules/.";
-import { LinkButtonType } from "@Common/type";
 
-type props = LinkButtonType;
-export const LinkButton: React.FC<props> = ({ url, type, content, onClick, refProps, id }) => {
+type LinkButtonType = {
+  url: string | undefined;
+  type: string;
+  content: string;
+  onClick?: React.MouseEventHandler<HTMLButtonElement> | undefined;
+  refProps?: RefObject<HTMLButtonElement>;
+  id?: string;
+};
+
+export const LinkButton: React.FC<LinkButtonType> = ({ url, type, content, onClick, refProps, id }) => {
   if (!(url && url !== ""))
     return (
       <Button size={type} onClick={onClick} ref={refProps} className={content === "로그아웃" ? "logout" : ""} data-id={id}>
