+++ client/src/Component/Organism/Chat/ChatListContainer.tsx
@@ -1,17 +1,12 @@
-import React from "react";
+import React, { RefObject } from "react";
 import { css } from "@emotion/react";
-import { ChatListContainerType } from "@Common/type";
 import { ChatImageContainer } from "./ChatImageContainer";
 import { ChatListDetail } from "./ChatListDetail";
 
-const ChatListContainerStyle = css`
-  width: 50%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  min-width: 520px;
-`;
+type ChatListContainerType = {
+  profileRef: RefObject<HTMLDivElement[]>;
+  onClick: (prev: any) => void;
+};
 
 export const ChatListContainer = ({ profileRef, onClick }: ChatListContainerType) => {
   return (
@@ -21,3 +16,12 @@ export const ChatListContainer = ({ profileRef, onClick }: ChatListContainerType
     </div>
   );
 };
+
+const ChatListContainerStyle = css`
+  width: 50%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  min-width: 520px;
+`;
