+++ client/src/Component/Template/Modal/InviteModal.tsx
@@ -4,7 +4,6 @@ import { useSetRecoilState } from "recoil";
 import { Button } from "@Atom/.";
 import { InputLabel } from "@Core/InputLabel";
 import { inviteTeam } from "@Common/api";
-import { TeamInfoType } from "@Common/type";
 import { errorState } from "@Recoil/Atom";
 import { teamState } from "@Recoil/TeamData";
 
@@ -33,15 +32,12 @@ export default function InviteModal({ handleFalseInvite }: { handleFalseInvite: 
   const clickInvite: MouseEventHandler = async () => {
     try {
       const inviteUserId = validationUserId({ userIdRef });
-      inviteTeam({ userId: inviteUserId })
-        .then(({ data }) => {
-          setTeamInfo((prev: TeamInfoType) => ({
-            ...prev,
-            member: [...(prev.member ?? []), data],
-          }));
-          handleFalseInvite();
-        })
-        .catch((e) => setErrorValue({ errorStr: "초대 실패했습니다", timeOut: 1000 }));
+      const { data } = await inviteTeam({ userId: inviteUserId });
+      setTeamInfo((prev) => ({
+        ...prev,
+        member: [...(prev.member ?? []), data],
+      }));
+      handleFalseInvite();
     } catch (e) {
       setErrorValue({ errorStr: e as string, timeOut: 1000 });
     }
